moved commands.d.ts to /lib/types
