+++ apps/bot/src/lib/types.d.ts
